==> scripts/04_BoundaryBased.Rmd <==
---
title: An R Markdown document converted from "04_BoundaryBased.ipynb"
output: html_document
---

# Boundary-based clustering

```{r}
library(readr)
library(dplyr, warn.conflicts = FALSE)
library(tidyr, warn.conflicts = FALSE)
library(purrr)
library(ggplot2)
```

```{r}
options(repr.plot.res = 300)
```

## Loading data

```{r}
load("../data/validclusmod.RData")
```

```{r}
print(validclusmod)
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        umapmod = map(
            sex,
            ~uwot::load_uwot(paste0("../data/ukb/umapmod_", .x))
        )
    )
print(validclusmod)
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        residtab = map(
            umapmod,
            ~map(
                1:.x$nn_index$ann$getNItems(),
                function(i){
                    .x$nn_index$ann$getItemsVector(i-1)
                }
            )
        ),
        residtab = map(residtab, do.call, what = rbind)
    )
print(validclusmod)
```

## Archetypes

```{r}
archmod <- validclusmod %>%
    transmute(
        sex,
        choosemod = map(
            residtab,
            ~archetypes::stepArchetypes(
                .x, 
                k = 3:20,
                nrep = 1,
                verbose = FALSE
            )
        )
    )
```

## Finding best model

```{r}
options(repr.plot.height = 3, repr.plot.width = 7)
archmod %>%
    mutate(
        choosemod = map(
            choosemod,
            ~{
                res <- unclass(.x)
                res <- map(res, unclass)
                res <- map(res, \(x) x[[1]])
                res <- map_dbl(res, "rss")
                data.frame(
                    Nclus = 3:20,
                    RSS = res
                )
            }
        )
    ) %>%
    unnest(choosemod) %>%
    ggplot(aes(Nclus, RSS)) +
    geom_line() +
    geom_point() +
    facet_wrap(~sex, scales = "free") +
    theme_bw()
```

## Selecting best model

```{r}
archbest <- validclusmod %>%
    transmute(
        sex,
        mod = map(
            residtab, 
            archetypes::archetypes, 
            k = 10
        )
    )
```

## Cluster centers

```{r}
options(repr.plot.height = 4, repr.plot.width = 8)
archbestcent <- archbest %>%
    transmute(
        sex,
        meandat = map(mod, ~.x$archetypes)
    )
print(archbestcent)
```

```{r}
options(repr.plot.height = 4, repr.plot.width = 10)
archbestcent %>%
    mutate(
        meandat = map(
            meandat,
            ~{
                res <- .x
                colnames(res) <- c("whr", "sbp", "dbp", "alt", "scr", "crp", "hdl", "tg", "ldl", "fg")
                data.frame(Cluster = paste0("Cluster", 1:nrow(res)), res)
            }
        )
    ) %>%
    unnest(meandat) %>%
    pivot_longer(-c(sex, Cluster), names_to = "trait", values_to = "residval") %>%
    mutate(trait = toupper(trait)) %>%
    ggplot(aes(residval, trait)) +
    geom_segment(aes(xend = 0, yend = trait)) +
    geom_vline(xintercept = 0, lty = "dashed") +
    geom_point() +
    facet_grid(sex ~ reorder(Cluster, as.numeric(gsub("Cluster", "", Cluster)))) +
    labs(x = "Residuals - SD units", y = NULL) +
    theme_bw()
```

## Cluster weights

```{r}
options(repr.plot.height = 2.5, repr.plot.width = 4.5)
archbest %>%
    transmute(
        sex,
        weightdat = map(
            mod,
            ~data.frame(
                Cluster = paste0("Cluster", 1:.x$k), 
                Weight = colSums(.x$alphas) / nrow(.x$alphas)
            )
        )
    ) %>%
    unnest(weightdat) %>%
    ggplot(aes(Weight, reorder(Cluster, desc(as.numeric(gsub("Cluster", "", Cluster)))))) +
    geom_col() +
    facet_wrap(~sex) +
    scale_x_continuous(labels = \(x) 100 * x) +
    theme_bw() +
    labs(x = "Weight (%)", y = NULL)
```

## Relative entropy

```{r}
archbest %>%
    transmute(
        sex,
        RelEnt = map_dbl(
            mod,
            ~{
                pp <- pmax(.x$alphas, .Machine$double.xmin)
                ent <- -sum(rowSums(pp * log(pp)))
                1 - ent / (nrow(pp) * log(ncol(pp)))
            }
        )
    )
```

## Projecting centers

```{r}
embeddf <- validclusmod %>%
    transmute(sex, umapmod = map(umapmod, ~data.frame(.x$embedding))) %>%
    unnest(umapmod)
head(embeddf)
```

```{r}
centembed <- validclusmod %>%
    inner_join(archbestcent, by = "sex") %>%
    transmute(
        sex,
        meandat = map2(
            meandat, umapmod,
            ~data.frame(uwot::umap_transform(.x, .y))
        )
    ) %>%
    unnest(meandat)
head(centembed)
```

```{r}
options(repr.plot.width = 5, repr.plot.height = 3)
embeddf %>%
    ggplot(aes(X1, X2)) +
    geom_point(shape = ".", alpha = .1) +
    geom_point(data = centembed, color = "red") +
    facet_wrap(~sex) +
    theme_bw() +
    labs(x = "UMAP1", y = "UMAP2")
```

## What if we used PCs instead

```{r}
pcdat <- validclusmod %>%
    mutate(
        pcd = map(residtab, prcomp)
    )
print(pcdat)
```

```{r}
archmodpc <- pcdat %>%
    transmute(
        sex,
        choosemod = map(
            pcd,
            ~archetypes::stepArchetypes(
                .x$x[,1:8], 
                k = 3:20,
                nrep = 1,
                verbose = FALSE
            )
        )
    )
```

```{r}
options(repr.plot.height = 3, repr.plot.width = 7)
archmodpc %>%
    mutate(
        choosemod = map(
            choosemod,
            ~{
                res <- unclass(.x)
                res <- map(res, unclass)
                res <- map(res, \(x) x[[1]])
                res <- map_dbl(res, "rss")
                data.frame(
                    Nclus = 3:20,
                    RSS = res
                )
            }
        )
    ) %>%
    unnest(choosemod) %>%
    ggplot(aes(Nclus, RSS)) +
    geom_line() +
    geom_point() +
    facet_wrap(~sex, scales = "free") +
    theme_bw()
```

```{r}
archbestpc <- pcdat %>%
    transmute(
        sex,
        mod = map(
            pcd,
            ~archetypes::archetypes(
                .x$x[,1:8], 
                k = 10
            )
        )
    )
```

```{r}
archbestpc %>%
    transmute(
        sex,
        RelEnt = map_dbl(
            mod,
            ~{
                pp <- pmax(.x$alphas, .Machine$double.xmin)
                ent <- -sum(rowSums(pp * log(pp)))
                1 - ent / (nrow(pp) * log(ncol(pp)))
            }
        ),
        RelEnt = round(RelEnt, 2)
    )
```
