==> scripts/09_NoDx.Rmd <==
---
title: An R Markdown document converted from "09_NoDx.ipynb"
output: html_document
---

```{r}
library(readr)
library(dplyr, warn.conflicts = FALSE)
library(tidyr, warn.conflicts = FALSE)
library(purrr)
library(ggplot2)
library(Matrix, warn.conflicts = FALSE)
library(uwot)
```

```{r}
options(repr.plot.res = 300)
```

```{r}
load("../data/validclusmod.RData")
```

```{r}
print(validclusmod)
```

```{r}
residtab <- rio::import("../data/ukb/residtab.RData")
map(residtab, head)
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        residtabs = map(
            sex,
            ~residtab[[.x]]
        )
    )
print(validclusmod)
```

```{r}
BIOMARKERS <- c("whr", "sbp", "dbp", "alt", "scr", "crp", "hdl", "tg", "ldl", "fg")
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        residtabs = map2(
            residtabs, clusmod,
            function(RESIDDF, CLUSMOD){
                RESDF <- select(RESIDDF, all_of(BIOMARKERS))
                pdfs <- mapply(
                    function(mu, covmat){ mvtnorm::dmvnorm(RESDF, mu, covmat) },
                    CLUSMOD$validclus_centers, 
                    CLUSMOD$validclus_covmats
                )
                L <- pdfs %*% diag(CLUSMOD$validclus_weights)
                colnames(L) <- paste0("prob", CLUSMOD$validclus_name)
                bind_cols(RESIDDF, data.frame(L / rowSums(L)))
            }
        )
    )
print(validclusmod)
```

```{r}
origprop <- validclusmod %>%
    transmute(
        sex,
        PropDF = map(
            residtabs,
            ~.x %>%
                summarise(across(starts_with("prob"), sum)) %>%
                pivot_longer(everything()) %>%
                mutate(N = sum(value), prop = value/N)
        )
    ) %>%
    unnest(PropDF)
origprop
```

```{r}
covar_dat <- read_tsv("../data/covar_dat.tsv", show_col_types = FALSE)
head(covar_dat)
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        dxdat = map(
            residtabs,
            inner_join,
            covar_dat,
            by = "eid"
        )
    )
print(validclusmod)
```

```{r}
DX <- c("HT", "CHD", "Stroke", "PAD", "CKD", "LiverFailure", "RA", "T2D", "T1D")
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        nodx = map(
            dxdat,
            ~.x %>%
                mutate(
                    anydx = rowSums(pick(all_of(DX))) != 0
                ) %>%
                filter(!anydx)
        )
    )
print(validclusmod)
```

```{r}
nodxprop <- validclusmod %>%
    transmute(
        sex,
        PropDF = map(
            nodx,
            ~.x %>%
                summarise(across(starts_with("prob"), sum)) %>%
                pivot_longer(everything()) %>%
                mutate(N = sum(value), prop = value/N)
        )
    ) %>%
    unnest(PropDF)
nodxprop
```

```{r}
cluscolmap <- validclusmod %>%
    transmute(
        clusmod = map(clusmod, select, Cluster = validclus_name, W = validclus_weights)
    ) %>%
    unnest(clusmod) %>%
    group_by(Cluster) %>%
    summarise(W2 = exp(sum(log((W))))) %>%
    arrange(-W2) %>%
    mutate(
        Cluster = factor(Cluster, levels = Cluster),
        ClusCol = scales::hue_pal()(n())
    )
cluscolmap
```

```{r}
options(repr.plot.width = 4, repr.plot.height = 3)
dxfreeprop <- bind_rows(
    All = origprop,
    `No Dx` = nodxprop,
    .id = "Subset"
) %>%
    mutate(Cluster = gsub("prob", "", name)) %>%
    select(-c(name, value, N)) %>%
    inner_join(cluscolmap, by = "Cluster") %>%
    expand_grid(plotrang = c("low", "high")) %>%
    mutate(
        Cluster = factor(Cluster, levels = cluscolmap$Cluster),
        prop = 100 * prop,
        propmin = case_when(
            plotrang == "low" ~ 0,
            prop < 10 ~ NaN, 
            .default = 60
        ),
        propmax = case_when(
            plotrang == "low" &  prop < 10 ~ prop,
            plotrang == "low" & prop > 10 ~ 10,
            plotrang == "high" & prop < 10 ~ NaN,
            plotrang == "high" & prop > 10 ~ prop
        )
    ) %>%
    ggplot(aes(Cluster, ymin = propmin, ymax = propmax)) +
    geom_linerange(
        aes(
            alpha = ifelse(Subset == "No Dx", .5, 1), 
            color = ClusCol,
            group = Subset
        ),
        position = position_dodge(width = .5),
        linewidth = 5,
        na.rm = TRUE
    ) +
    scale_color_identity() +
    scale_alpha_identity(
        breaks = c(1, .5),
        labels = c("All", "Diagnosis-free"),
        name = NULL,
        guide = guide_legend()
    ) +
    facet_grid(
        plotrang ~ sex,
        scales = "free",
        space = "free_x"
    ) +
    ggh4x::force_panelsizes(rows = c(.35, .65)) +
    ggh4x::scale_y_facet(
        plotrang == "low", 
        expand = c(0,0)
    ) +
    ggh4x::scale_y_facet(
        plotrang == "high", breaks = c(70, 80), 
        expand = expansion(mult = c(0,0.1)),
        #labels = \(x)ifelse(x == 60, "", x)
    ) +
    theme_bw() +
    theme(
        strip.text.y = element_blank(),
        strip.background.y = element_blank(),
        legend.position = "top"
    ) +
    labs(x = NULL, y = "%")
dxfreeprop
```

```{r}
saveRDS(dxfreeprop, file = "../data/dxfreeprop.rds")
```

```{r}
strat_dat <- rio::import("../data/ukb/strat_dat.RData")
map(strat_dat, head)
```

```{r}
MetSDat <- map(
    strat_dat,
    inner_join,
    covar_dat,
    by = "eid"
) %>%
    map(
        transmute,
        eid,
        MetS1 = T2D == 1 | fg > 5.6,
        MetS2 = 1 * (
            (bmi >= 30) |
            (sex == "Male" & whr > .9) |
            (sex == "Female" & whr > .85)
        ),
        MetS3 = 1 * (
            (tg >= 1.7) |
            (sex == "Male" & hdl < 1.03) |
            (sex == "Female" & hdl < 1.3)
        ),
        MetS4 = 1 * (
            (sbp >= 140 & dbp >= 90) |
            AntiHT == 1
        ),
        MetS = 1 * (MetS1 & ((MetS2 + MetS3 + MetS4) >= 2)),
        across(matches("[0-9]$"), \(x) NULL)
    )
map(MetSDat, head)
```

```{r}
map(MetSDat, count, MetS)
```

```{r}
validclusmod <- validclusmod %>%
    mutate(
        metstabs = map(sex, ~MetSDat[[.x]]),
        metstabs = map2(
            residtabs, metstabs,
            inner_join, by = "eid"
        )
    )
print(validclusmod)
```

```{r}
validclusmod %>%
    transmute(
        sex,
        metstabs = map(
            metstabs,
            ~.x %>%
                summarise(
                    across(starts_with("prob"), \(x) sum(x * MetS))
                ) %>%
                pivot_longer(everything()) %>%
                mutate(N = sum(value), prop = value/N)
        )
    ) %>%
    unnest(metstabs)
```

```{r}
metsres <- validclusmod %>%
    transmute(
        sex,
        metstabs = map(
            metstabs,
            ~.x %>%
                select(
                    eid, MetS, starts_with("prob")
                ) %>%
                pivot_longer(
                    starts_with("prob"), 
                    names_to = "Cluster"
                ) %>%
                group_by(Cluster) %>%
                summarise(
                    N = sum(value),
                    Ncases = sum(MetS * value),
                    Prop = 100*Ncases/N
                ) %>%
                mutate(
                    Cluster = factor(
                        gsub("prob", "", Cluster), 
                        levels = cluscolmap$Cluster
                    ),
                    Ncontrols = N - Ncases
                )
        )
    ) %>%
    unnest(metstabs)
metsres
```

```{r}
metsest <- metsres %>%
    nest(Dat = -sex) %>%
    mutate(
        mod = map(
            Dat,
            ~glm(cbind(round(Ncases), round(Ncontrols)) ~ Cluster, data = .x, family = "binomial")
        ),
        mod = map(mod, broom::tidy),
        mod = map(mod, mutate, term = gsub("Cluster", "", term)),
        Dat = map2(Dat, mod, left_join, by = c("Cluster" = "term")),
        mod = NULL
    ) %>%
    unnest(Dat) %>%
    mutate(
        OR = exp(estimate),
        LOWERCI = exp(estimate - qnorm(1 - .05/2)*std.error),
        UPPERCI = exp(estimate + qnorm(1 - .05/2)*std.error)
    )
metsest
```

```{r}
options(repr.plot.width = 3, repr.plot.height = 3)
metsest %>%
    drop_na %>%
    mutate(
        bigeff = OR > 10, xint = ifelse(bigeff, NA, 1),
        Cluster = factor(Cluster, levels = rev(cluscolmap$Cluster))
    ) %>%
    ggplot(aes(OR, Cluster)) +
    geom_vline(aes(xintercept = xint), lty = "dashed", na.rm = TRUE) +
    geom_linerange(
        aes(xmin = LOWERCI, xmax = UPPERCI)
    ) +
    geom_point() +
    facet_grid(
        sex ~ bigeff, 
        scales = "free",
        space = "free_y"
    ) +
    ggh4x::force_panelsizes(cols = c(.7, .3)) +
    ggh4x::scale_x_facet(bigeff, n.breaks = 3) +
    theme_bw() +
    theme(
        strip.text.x = element_blank(),
        strip.background.x = element_blank()
    ) +
    labs(x = "OR relative to BC", y = NULL)
```
